wip(flows): Search for flow modules only in the directory of the core. Working towards https://github.com/IObundle/py2hwsw/issues/167
With this commit, flow modules (blocks that have "dest_dir" different
from "hardware/src/") will only be searched for in the directory of the
core that calls them (their instantiator).

Currently has issues with automatic memory wrapper, because its superblocks are taken from iob_system, and are not contained inside the iob_memwrapper folder.